Replace jingo with django-jinja
