+++ oneanddone/base/tests/test_helpers.py
@@ -1,35 +1,86 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
-from django.test.utils import override_settings
 
-from mock import patch
+import urlparse
+
+from django.http import QueryDict
+from mock import Mock
 from nose.tools import eq_
 
-from oneanddone.base.helpers import less_css
 from oneanddone.base.tests import TestCase
+from oneanddone.base.templatetags.jinja_helpers import page_url, buglinkify
 
 
-class LessCssTests(TestCase):
-    @override_settings(TEMPLATE_DEBUG=False)
-    def test_template_debug_false_call_css(self):
-        """If TEMPLATE_DEBUG is false, call jingo_minify.helpers.css."""
-        with patch('oneanddone.base.helpers.css') as mock_css:
-            eq_(less_css('bundle'), mock_css.return_value)
-            mock_css.assert_called_with('bundle')
-
-    @override_settings(TEMPLATE_DEBUG=True)
-    def test_template_debug_true_less_tags(self):
+class PageUrlTests(TestCase):
+    def test_basic(self):
         """
-        If TEMPLATE_DEBUG is true, return a set of link tags with
-        stylesheet/less as the rel attribute.
+        page_url should return a relative link to the current page,
+        preserving the GET arguments from the given request, and adding
+        a page parameter for the given page.
         """
-        with patch('oneanddone.base.helpers.get_css_urls') as get_css_urls:
-            get_css_urls.return_value = ['foo', 'bar']
-            output = less_css('bundle')
-            get_css_urls.assert_called_with('bundle')
 
-        self.assertHTMLEqual(output, """
-            <link rel="stylesheet/less" media="screen,projection,tv" href="foo" />
-            <link rel="stylesheet/less" media="screen,projection,tv" href="bar" />
-        """)
+        request = Mock(GET=QueryDict('foo=bar&baz=5'))
+        url = urlparse.urlsplit(page_url(request, 4))
+        args = urlparse.parse_qs(url.query)
+        eq_(args, {'foo': ['bar'], 'baz': ['5'], 'page': ['4']})
+
+    def test_existing_page_arg(self):
+        """
+        If the current page already has a page GET argument, override
+        it.
+        """
+        request = Mock(GET=QueryDict('foo=bar&page=5'))
+        url = urlparse.urlsplit(page_url(request, 4))
+        args = urlparse.parse_qs(url.query)
+        eq_(args, {'foo': ['bar'], 'page': ['4']})
+
+    def test_repeats(self):
+        """
+        GET parameters with multiple values should have all their
+        values preserved
+        """
+        request = Mock(GET=QueryDict('foo=bar&baz=5&foo=ok'))
+        url = urlparse.urlsplit(page_url(request, 4))
+        args = urlparse.parse_qs(url.query)
+        eq_(args, {'foo': ['bar', 'ok'], 'baz': ['5'], 'page': ['4']})
+
+
+class BuglinkifyTests(TestCase):
+
+    def setUp(self):
+        self.bugzilla_url_prefix = 'https://bugzilla.mozilla.org/show_bug.cgi?id='
+
+    def test_with_bug_lcase(self):
+        """
+        buglinkify should linkify a lowercase bug in a string.
+        """
+        name = 'Test this string with bug 12345'
+        eq_(buglinkify(name),
+            'Test this string with <a href="%s12345">bug 12345</a>' %
+            self.bugzilla_url_prefix)
+
+    def test_with_bug_ucase(self):
+        """
+        buglinkify should linkify a capitalized bug in a string.
+        """
+        name = 'Test this string with Bug 12345'
+        eq_(buglinkify(name),
+            'Test this string with <a href="%s12345">Bug 12345</a>' %
+            self.bugzilla_url_prefix)
+
+    def test_without_bug(self):
+        """
+        buglinkify should leave string intact without a bug.
+        """
+        name = 'Test this string with boog 12345'
+        eq_(buglinkify(name), name)
+
+    def test_with_extra_numbers(self):
+        """
+        buglinkify should only linkify the bug.
+        """
+        name = 'Test this 12345 string with Bug 12345 and 12345'
+        eq_(buglinkify(name),
+            'Test this 12345 string with <a href="%s12345">Bug 12345</a> and 12345' %
+            self.bugzilla_url_prefix)
