+++ oneanddone/settings/base.py
@@ -36,9 +36,6 @@ DEBUG = config('DJANGO_DEBUG', default=False, cast=bool)
 
 DEV = config('DEV', default=DEBUG, cast=bool)
 
-TEMPLATE_DEBUG = config('DEBUG', default=DEBUG, cast=bool)
-
-
 ALLOWED_HOSTS = config('ALLOWED_HOSTS', default='*', cast=Csv())
 
 HMAC_KEYS = {
@@ -80,8 +77,9 @@ INSTALLED_APPS = [
     'commonware.response.cookies',
     'django_ace',
     'django_browserid',
+    'django_jinja',
+    'django_jinja.contrib._humanize',  # Adds django humanize filters
     'django_nose',
-    'jingo_minify',
     'rest_framework',
     'rest_framework.authtoken',
     'tower',
@@ -90,18 +88,19 @@ INSTALLED_APPS = [
 
 MIDDLEWARE_CLASSES = (
     'sslify.middleware.SSLifyMiddleware',
-    'oneanddone.base.middleware.LocaleURLMiddleware',
     'django.middleware.common.CommonMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'session_csrf.CsrfMiddleware',  # Must be after auth middleware.
     'django.contrib.messages.middleware.MessageMiddleware',
     'commonware.middleware.FrameOptionsHeader',
+    'django.middleware.clickjacking.XFrameOptionsMiddleware',
+    'csp.middleware.CSPMiddleware',
     'oneanddone.base.middleware.TimezoneMiddleware',
     'oneanddone.base.middleware.ClosedTaskNotificationMiddleware',
 )
 
-TEMPLATE_CONTEXT_PROCESSORS = (
+CONTEXT_PROCESSORS = (
     'django.contrib.auth.context_processors.auth',
     'django.core.context_processors.debug',
     'django.core.context_processors.media',
@@ -112,14 +111,55 @@ TEMPLATE_CONTEXT_PROCESSORS = (
     'oneanddone.base.context_processors.globals',
 )
 
-TEMPLATE_DIRS = (
-    path('templates'),
-)
+TEMPLATES = [
+    {
+        'BACKEND': 'django_jinja.backend.Jinja2',
+        'DIRS': [],
+        'APP_DIRS': True,
+        'OPTIONS': {
+            # Use jinja2/ for jinja templates
+            'app_dirname': 'jinja2',
+            # Don't figure out which template loader to use based on
+            # file extension
+            'match_extension': '',
+            'newstyle_gettext': True,
+            'context_processors': CONTEXT_PROCESSORS,
+            'undefined': 'jinja2.Undefined',
+            'extensions': [
+                'jinja2.ext.do',
+                'jinja2.ext.loopcontrols',
+                'jinja2.ext.with_',
+                'jinja2.ext.i18n',
+                'jinja2.ext.autoescape',
+                'django_jinja.builtins.extensions.CsrfExtension',
+                'django_jinja.builtins.extensions.CacheExtension',
+                'django_jinja.builtins.extensions.TimezoneExtension',
+                'django_jinja.builtins.extensions.UrlsExtension',
+                'django_jinja.builtins.extensions.StaticFilesExtension',
+                'django_jinja.builtins.extensions.DjangoFiltersExtension',
+            ],
+            'globals': {
+                'browserid_info': 'django_browserid.helpers.browserid_info',
+                'browserid_login': 'django_browserid.helpers.browserid_login',
+                'browserid_logout': 'django_browserid.helpers.browserid_logout'
+            }
+        }
+    },
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'DIRS': [],
+        'APP_DIRS': True,
+        'OPTIONS': {
+            'debug': DEBUG,
+            'context_processors': CONTEXT_PROCESSORS
+        }
+    },
+]
 
-AUTHENTICATION_BACKENDS = (
-    'django.contrib.auth.backends.ModelBackend',
+AUTHENTICATION_BACKENDS = [
     'django_browserid.auth.BrowserIDBackend',
-)
+    'django.contrib.auth.backends.ModelBackend',
+]
 
 # Sessions
 #
@@ -206,12 +246,6 @@ MEDIA_URL = config('MEDIA_URL', default='/media/')
 
 SESSION_COOKIE_SECURE = config('SESSION_COOKIE_SECURE', default=not DEBUG, cast=bool)
 
-TEMPLATE_LOADERS = (
-    'jingo.Loader',
-    'django.template.loaders.filesystem.Loader',
-    'django.template.loaders.app_directories.Loader',
-)
-
 # Django-CSP
 CSP_DEFAULT_SRC = (
     "'self'",
@@ -248,65 +282,6 @@ SESSION_COOKIE_SECURE = not DEBUG
 
 # Third-party Library Settings
 ##############################################################################
-
-
-def JINJA_CONFIG():
-    config = {
-        'extensions': [
-            'tower.template.i18n',
-            'jinja2.ext.do',
-            'jinja2.ext.with_',
-            'jinja2.ext.loopcontrols'
-        ],
-        'finalize': lambda x: x if x is not None else ''
-    }
-    return config
-
-# Because Jinja2 is the default template loader, add any non-Jinja templated
-# apps here:
-JINGO_EXCLUDE_APPS = [
-    'admin',
-    'registration',
-    'browserid',
-    'memcached',
-    'rest_framework',
-]
-
-# Bundles is a dictionary of two dictionaries, css and js, which list css files
-# and js files that can be bundled together by the minify app.
-MINIFY_BUNDLES = {
-    'css': {
-        'base': (
-            'browserid/persona-buttons.css',
-            'css/sandstone/sandstone-resp.less',
-            'css/one-and-done.less',
-            'css/slider.css',
-            'css/smoothness/jquery-ui-1.10.4.custom.css',
-            'css/datatables/jquery.dataTables.css'
-        ),
-    },
-    'js': {
-        'base': (
-            'js/libs/jquery-2.0.3.min.js',
-            'browserid/api.js',
-            'browserid/browserid.js',
-            'js/site.js',
-            'js/slider.js',
-            'js/libs/jquery-ui-1.10.4.custom.js',
-            'js/libs/jquery.dataTables.js'
-        ),
-    }
-}
-
-# Use staticfiles loaders for finding resources for minification.
-JINGO_MINIFY_USE_STATIC = True
-
-# Path to Java. Used for compress_assets.
-JAVA_BIN = '/usr/bin/java'
-
-# Do not preprocess LESS files.
-LESS_PREPROCESS = False
-LESS_BIN = config('LESSC_BIN', default='lessc')
 
 # Testing configuration.
 NOSE_ARGS = ['--logging-clear-handlers', '--logging-filter=-factory,-south,-django.db']
