+++ oneanddone/tasks/jinja2/tasks/submit_admin_verification_response_form.html
@@ -3,13 +3,13 @@
  # file, You can obtain one at http://mozilla.org/MPL/2.0/. #}
 
 <form method="post" class="feedback-form">
-  <input type="hidden" id="pk" name="pk" value="{{ task.id }}">
-  {{ csrf() }}
+  <input type="hidden" id="pk" name="pk" value="{{ taskattempt.id }}">
+  {% csrf_token %}
   <p>
-    {{ verification_form.content }}
-    {{ verification_form.content.errors }}
+    {{ form.content }}
+    {{ form.content.errors }}
   </p>
   <div class="right">
-    <button type="submit" class="button">{{ _('Submit for verification') }}</button>
+    <button type="submit" class="button">{{ _('Submit feedback to user') }}</button>
   </div>
 </form>
