+++ oneanddone/users/jinja2/users/profile/delete.html
@@ -9,7 +9,7 @@
 <main class="billboard content-container">
 <h1>{{_('Are you sure you want to delete your profile ?')}}</h1>
   <form method="post" class="edit-profile">
-    {{ csrf() }}
+    {% csrf_token %}
     <p>{{_('Deleting your profile will delete all your personal data and your task data.')}}</p>
     <div class="actions-container">
         <a href="{{ url('base.home') }}" id="cancel-button" class="button">{{ _('Cancel') }}</a>
